add interfaces, rename implementation classes
