+++ src/main/java/wawer/kamil/App.java
@@ -1,7 +1,6 @@
 package wawer.kamil;
 
-import wawer.kamil.accountService.AccountService;
-import wawer.kamil.accountService.AccountValidators;
+import wawer.kamil.accountService.AccountServiceImpl;
 import wawer.kamil.model.Account;
 import wawer.kamil.model.AccountList;
 import wawer.kamil.model.Parser;
@@ -14,12 +13,12 @@ public class App {
     public static void main(String[] args) throws IOException {
 
         Parser parser = new Parser();
-        AccountService service = new AccountService();
+        AccountServiceImpl service = new AccountServiceImpl();
         AccountList list = parser.readFromXML();
 
         List<Account> listOfAccounts = list.getAccountList();
         List<Account> listOfValidatedAccounts = service.validateEverySingleAccount(listOfAccounts);
-        List<Account> listOfSortedAccounts = service.sotrValidatedAccountList(listOfValidatedAccounts);
+        List<Account> listOfSortedAccounts = service.sortValidatedAccountList(listOfValidatedAccounts);
         parser.writeToXML(listOfSortedAccounts);
 
         System.out.println("\n Lista na wejściu: " + listOfAccounts + "\n");
