+++ src/test/java/AccountValidatorsTest.java
@@ -1,5 +1,5 @@
 /*
-import wawer.kamil.accountService.AccountValidators;
+import wawer.kamil.accountService.AccountValidatorsImpl;
 import wawer.kamil.model.Account;
 
 import java.math.BigDecimal;
@@ -7,11 +7,11 @@ import java.time.LocalDate;
 
 public class AccountValidatorsTest {
 
-    private AccountValidators validators;
+    private AccountValidatorsImpl validators;
     private Account account;
 
     public AccountValidatorsTest() {
-        this.validators = new AccountValidators();
+        this.validators = new AccountValidatorsImpl();
         this.account = new Account("PL61109010140000071219812870", "name4", "PLN", BigDecimal.ZERO, LocalDate.parse("2029-10-11"));
     }
